server/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
pandas>=2.0.0
geopandas>=0.13.0
flask-sqlalchemy>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn==0.24.2
fiona==1.8.20
shapely==1.7.1
gunicorn==20.1.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
requests==2.26.0

